requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
gunicorn>=20.1.0